# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API pro evidenci denních pracovních záznamů - Kácení a Sečení
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Seznam uživatelů (pouze admin)
      responses:
        "200":
          description: Seznam uživatelů
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Vytvořit uživatele (pouze admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Uživatel vytvořen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Získat uživatele
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Uživatel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Nenalezeno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Upravit uživatele
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Uživatel upraven
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Smazat uživatele (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Uživatel smazán
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, user, employee, manager]
        workerId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - username
        - fullName
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, user, employee, manager]
        workerId:
          type: ["integer", "null"]
      required:
        - username
        - password
        - fullName
        - role
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          enum: [admin, user, employee, manager, null]
        workerId:
          type: ["integer", "null"]
        isActive:
          type: ["boolean", "null"]
        password:
          type: ["string", "null"]
...